Move ConnectionSecrets into tls12 module
